Modularize deprecation related functions
And udpate the affected code accordingly.
