+++ src/misc.js
@@ -1,4 +1,5 @@
 const defaultTimeoutInMs = require("./constants_and_utils/constants").DEFAULT_TIMEOUT_IN_MS;
+const deprecation = require("./constants_and_utils/deprecation");
 const messageBuilder = require("./constants_and_utils/messageBuilder");
 const utils = require("./constants_and_utils/utils");
 const waiters = require("./waiters");
@@ -6,7 +7,7 @@ const waiters = require("./waiters");
 const getBodyElementFromCurrentBrowserOrBrowserInstance = function(browserInstance) {
   const cssSelector = "body";
 
-  utils.obsoleteFunction("getBodyElementFromCurrentBrowserOrBrowserInstance");
+  deprecation.obsoleteFunction("getBodyElementFromCurrentBrowserOrBrowserInstance");
   if (browserInstance) {
     return browserInstance.element(by.css(cssSelector));
   }
@@ -14,7 +15,7 @@ const getBodyElementFromCurrentBrowserOrBrowserInstance = function(browserInstan
 };
 
 const openNewBrowserInTheSamePage = function(browser = requiredParam(openNewBrowserInTheSamePage, "browser")) {
-  utils.obsoleteFunction("openNewBrowserInTheSamePage");
+  deprecation.obsoleteFunction("openNewBrowserInTheSamePage");
   return browser.forkNewDriverInstance(true);
 };
 
@@ -26,7 +27,7 @@ const scrollToElementWhenVisible = function(
   htmlElement = utils.requiredParam(scrollToElementWhenVisible),
   timeoutInMilliseconds = utils.timeout.timeoutInMilliseconds
 ) {
-  utils.replaceObsoleteFunction("scrollToElementWhenVisible", "scrollToElement");
+  deprecation.replaceObsoleteFunction("scrollToElementWhenVisible", "scrollToElement");
   this.scrollToElement(htmlElement, timeoutInMilliseconds);
 };
